+++ Frontend/android-native/app/src/driver/nammaYatriPartner/assets/juspay/juspay_assets.json
@@ -609,6 +609,7 @@
       "ny_ic_money_outline" : true,
       "ny_ic_more_horizontal_grey" : true,
       "ny_ic_more_horizontal_white" : true,
+      "ny_ic_more_stops_left" : true,
       "ny_ic_more_vertical" : true,
       "ny_ic_mysuru_map" : true,
       "ny_ic_namma_yatri_driver_logo" : true,
@@ -811,6 +812,7 @@
       "ny_ic_star_white_border" : true,
       "ny_ic_start_record" : true,
       "ny_ic_static_record" : true,
+      "ny_ic_stop_grey" : true,
       "ny_ic_stop_record" : true,
       "ny_ic_sub_save_more" : true,
       "ny_ic_submit_issue_arrow" : true,
