+++ Frontend/android-native/mobility-app/src/main/java/in/juspay/mobility/app/NotificationUtils.java
@@ -288,6 +288,7 @@ public class NotificationUtils {
                     sheetData.putInt("driverStepFeeWithCurrency", driverStepFeeWithCurrency.optInt("amount", negotiationUnit));
                     sheetData.putDouble("parkingCharge", entity_payload.optDouble("parkingCharge", 0.0));
                     sheetData.putBoolean("isFavourite", entity_payload.has("isFavourite") && entity_payload.optBoolean("isFavourite", false));
+                    sheetData.putInt("middleStopCount", entity_payload.optInt("middleStopCount", 0));
                     expiryTime = entity_payload.getString("searchRequestValidTill");
                     searchRequestId = entity_payload.getString("searchRequestId");
                     System.out.println(entity_payload);
