+++ Frontend/android-native/mobility-app/src/main/java/in/juspay/mobility/app/RideRequestActivity.java
@@ -120,7 +120,7 @@ public class RideRequestActivity extends AppCompatActivity {
             String rideDuration = String.format("%02d:%02d Hr", rideRequestBundle.getInt("rideDuration") / 3600 ,( rideRequestBundle.getInt("rideDuration") % 3600 ) / 60);
             String rideDistance = String.format("%d km", rideRequestBundle.getInt("rideDistance") / 1000);
             String notificationSource= rideRequestBundle.getString("notificationSource");
-                    
+            int stops = rideRequestBundle.getInt("middleStopCount", 0);
             SheetModel sheetModel = new SheetModel((df.format(distanceToPickup / 1000)),
                     distanceTobeCovered,
                     tollCharges,
@@ -166,7 +166,8 @@ public class RideRequestActivity extends AppCompatActivity {
                     rideRequestBundle.getBoolean("isThirdPartyBooking"),
                     rideRequestBundle.getBoolean("isFavourite"),
                     rideRequestBundle.getDouble("parkingCharge"),
-                    getCurrTime
+                    getCurrTime,
+                    stops
                     );
             sheetArrayList.add(sheetModel);
             sheetAdapter.updateSheetList(sheetArrayList);
@@ -203,6 +204,8 @@ public class RideRequestActivity extends AppCompatActivity {
                         ColorStateList.valueOf(getColor(R.color.Black900)) :
                         ColorStateList.valueOf(getColor(R.color.green900)));
                 holder.rideTypeTag.setVisibility(showVariant ? View.VISIBLE : View.GONE);
+                holder.stopsTag.setVisibility(model.getStops() > 0 ? View.VISIBLE : View.GONE);
+                holder.stopsTagText.setText(getString(R.string.stops, model.getStops()));
             } else {
                 holder.tagsBlock.setVisibility(View.GONE);
             }
@@ -271,6 +274,8 @@ public class RideRequestActivity extends AppCompatActivity {
 
             holder.textIncPrice.setText(String.valueOf(model.getNegotiationUnit()));
             holder.textDecPrice.setText(String.valueOf(model.getNegotiationUnit()));
+            holder.stopsInfo.setText(getString(model.getTollCharges() > 0 ? R.string.stops : R.string.stops_info, model.getStops()));
+            holder.stopsInfo.setVisibility(model.getStops() > 0? View.VISIBLE : View.GONE);
             if(model.getSourcePinCode() != null &&  model.getSourcePinCode().trim().length()>0){
                 holder.sourcePinCode.setText(model.getSourcePinCode().trim());
                 holder.sourcePinCode.setVisibility(View.VISIBLE);
