+++ Frontend/android-native/mobility-app/src/main/res/drawable/ic_stops_bg.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <corners android:radius="13dp" />
+    <stroke
+        android:width="1dp"
+        android:color="@color/Black700" />
+    <solid android:color="@color/Black700" />
+</shape>
