+++ Frontend/android-native/mobility-app/src/main/res/values-bn/strings.xml
@@ -77,4 +77,6 @@
     <string name="third_party_booking">"তৃতীয় পক্ষের বুকিং"</string>
     <string name="parking_charges">%1$.0f₹ পার্কিং</string>
     <string name="includes">অন্তর্ভুক্ত</string>
+    <string name="stops_info">%1$d স্টপস (স্টপ চার্জ সহ)</string>
+    <string name="stops">%1$d স্টপস</string>
 </resources>
