+++ Frontend/android-native/mobility-app/src/main/res/values-kn/strings.xml
@@ -82,4 +82,6 @@
     <string name="third_party_booking">"ತೃತೀಯ ಪಕ್ಷದ ಬುಕಿಂಗ್"</string>
     <string name="parking_charges">%1$.0f₹ ಪಾರ್ಕಿಂಗ್</string>
     <string name="includes">ಒಳಗೊಂಡಿದೆ</string>
+    <string name="stops_info">%1$d ನಿಲ್ದಾಣಗಳು (ನಿಲ್ದಾಣ ಶುಲ್ಕ ಸೇರಿದಂತೆ)</string>
+    <string name="stops">%1$d ನಿಲ್ದಾಣಗಳು</string>
 </resources>
