+++ Frontend/android-native/mobility-app/src/main/res/values-ta/strings.xml
@@ -83,4 +83,6 @@
     <string name="third_party_booking">"மூன்றாம் தரப்பு முன்பதிவு"</string>
     <string name="parking_charges">%1$.0f₹ பார்க்கிங்</string>
     <string name="includes">அடங்கும்</string>
+    <string name="stops_info">%1$d நிறுத்தங்கள் (நிறுத்த கட்டணத்துடன்)</string>
+    <string name="stops">%1$d நிறுத்தங்கள்</string>
 </resources>
