+++ Frontend/android-native/mobility-app/src/main/res/values/strings.xml
@@ -92,4 +92,6 @@
     <string name="third_party_booking">"Third Party Booking"</string>
     <string name="parking_charges">%1$.0f₹ Parking</string>
     <string name="includes">Includes</string>
+    <string name="stops_info">%1$d Stops (incl. stop charge)</string>
+    <string name="stops">%1$d Stops</string>
 </resources>
