+++ Frontend/android-native/mobility-customer/src/main/java/in/juspay/mobility/customer/MobilityCustomerBridge.java
@@ -71,7 +71,7 @@ public class MobilityCustomerBridge extends MobilityCommonBridge {
     public enum MapMode {
         NORMAL, SPECIAL_ZONE, HOTSPOT
     }
-    private Integer debounceAnimateCameraCounter;
+
 
     public MobilityCustomerBridge(BridgeComponents bridgeComponents) {
         super(bridgeComponents);
@@ -81,217 +81,9 @@ public class MobilityCustomerBridge extends MobilityCommonBridge {
     // endregion
 
     //region Maps
-    @JavascriptInterface
-    public String isCoordOnPath(String json, double currLat, double currLng, int speed) throws JSONException {
-        LatLng currPoint = new LatLng(currLat, currLng);
-        ArrayList<LatLng> path = new ArrayList<>();
-        JSONObject jsonObject = new JSONObject(json);
-        int distanceRemaining;
-        JSONArray coordinates = jsonObject.getJSONArray("points");
-        int eta;
-        int resultIndex;
-        JSONObject result = new JSONObject();
-        for (int i = coordinates.length() - 1; i >= 0; i--) {
-            JSONObject coordinate = (JSONObject) coordinates.get(i);
-            double lng = coordinate.getDouble("lng");
-            double lat = coordinate.getDouble("lat");
-            LatLng tempPoints = new LatLng(lat, lng);
-            path.add(tempPoints);
-        }
-        if (path.size() == 0) {
-            result.put("points", new JSONArray());
-            result.put("eta", 0);
-            result.put("distance", 0);
-            result.put("isInPath", false);
-            return result.toString();
-        }
-        double locationOnPathThres = Double.parseDouble(getKeysInSharedPref("ACCURACY_THRESHOLD").equals("__failed") ? "30.0" : getKeysInSharedPref("ACCURACY_THRESHOLD"));
-        resultIndex = PolyUtil.locationIndexOnEdgeOrPath(currPoint, path, PolyUtil.isClosedPolygon(path), true, locationOnPathThres);
-        if (resultIndex == -1) {
-            result.put("points", coordinates);
-            result.put("eta", 0);
-            result.put("distance", 0);
-            result.put("isInPath", false);
-        } else if (resultIndex == (path.size() - 2) || resultIndex == (path.size() - 1)) {
-            distanceRemaining = (int) SphericalUtil.computeLength(path);
-            eta = distanceRemaining / speed;
-            result.put("points", coordinates);
-            result.put("eta", eta);
-            result.put("distance", distanceRemaining);
-            result.put("isInPath", true);
-        } else if (resultIndex == 0) {
-            path.clear();
-            result.put("points", new JSONArray());
-            result.put("eta", 0);
-            result.put("distance", 0);
-            result.put("isInPath", true);
-        } else {
-            path.subList(resultIndex + 2, path.size()).clear();
-            distanceRemaining = (int) SphericalUtil.computeLength(path);
-            eta = distanceRemaining / speed;
-            JSONArray remainingPoints = new JSONArray();
-            for (int i = path.size() - 1; i >= 0; i--) {
-                LatLng point = path.get(i);
-                JSONObject tempPoints = new JSONObject();
-                tempPoints.put("lat", point.latitude);
-                tempPoints.put("lng", point.longitude);
-                remainingPoints.put(tempPoints);
-            }
-            result.put("points", remainingPoints);
-            result.put("eta", eta);
-            result.put("distance", distanceRemaining);
-            result.put("isInPath", true);
-        }
-        return result.toString();
-    }
 
-    @JavascriptInterface
-    public void updateRoute(String _payload) {
-        ExecutorManager.runOnMainThread(() -> {
-            try {
-                JSONObject payload = new JSONObject(_payload);
-                String pureScriptID = payload.optString("pureScriptID","");
-                GoogleMap gMap = pureScriptID.isEmpty() ? googleMap : googleMapInstance.get(pureScriptID);
-                if (gMap != null) {
-                    try {
-                    MapRemoteConfig mapRemoteConfig = getMapRemoteConfig();
-                    float zoomLevel;
-                    try {
-                        zoomLevel = (float) mapRemoteConfig.zoomLevel;
-                    }catch (Exception e){
-                        zoomLevel = 20.0f;
-                    }
-                    String json = payload.optString("json", "");
-                    JSONObject destMarkerConfig = payload.optJSONObject("destMarkerConfig");
-                    String dest = (destMarkerConfig != null) ? destMarkerConfig.optString("pointerIcon", "") : "";
-                    String eta = payload.optString("eta", "");
-                    String locationName = payload.optString("locationName" , "");
-                    String src = payload.optString("srcMarker", "");
-                    String specialLocation = payload.optString("specialLocation", "");
-                    String polylineKey = payload.optString("polylineKey", "DEFAULT");
-                    JSONObject destMarkerActionImageConfig = (destMarkerConfig != null) ? destMarkerConfig.optJSONObject("actionImage") : null;
-                    JSONObject specialLocationObject = new JSONObject(specialLocation);
-                    int dashUnit = specialLocationObject.optInt("dashUnit", 1);
-                    int gapUnit = specialLocationObject.optInt("gapUnit", 0);
-                    boolean autoZoom = payload.optBoolean("autoZoom", true);
-                    ArrayList<LatLng> path = new ArrayList<>();
-                    JSONObject jsonObject = new JSONObject(json);
-                    JSONArray coordinates = jsonObject.getJSONArray("points");
-                    for (int i = coordinates.length() - 1; i >= 0; i--) {
-                        JSONObject coordinate = (JSONObject) coordinates.get(i);
-                        double lng = coordinate.getDouble("lng");
-                        double lat = coordinate.getDouble("lat");
-                        LatLng tempPoint = new LatLng(lat, lng);
-                        path.add(tempPoint);
-                    }
-                    Marker currMarker = (Marker) markers.get(src);
-                    if (currMarker != null)
-                        {
-                            currMarker.setTitle("Vehicle Icon On Map");
-                            currMarker.hideInfoWindow();
-                        }
-                    Marker destMarker = (Marker) markers.get(dest);
-                    String destinationSpecialTagIcon = specialLocationObject.getString("destSpecialTagIcon");
-                    MarkerConfig markerConfig = new MarkerConfig();
-                    markerConfig.locationName(locationName);
-                    markerConfig.setLabelImage(destinationSpecialTagIcon);
-                    if (destMarkerActionImageConfig != null) markerConfig.setMarkerActionImageConfig(destMarkerActionImageConfig);
-                    if (destMarker != null)
-                    {
-                        destMarker.setIcon((BitmapDescriptorFactory.fromBitmap(getMarkerBitmapFromView(dest, false, MarkerType.NORMAL_MARKER, markerConfig))));
-                        destMarker.setTitle("Driver is " + eta);
-                    }
-                    PolylineDataPoints polylineDataPoints = getPolyLineDataByMapInstance(pureScriptID,polylineKey);
-                    Polyline polyline = getPolyLine(false, polylineDataPoints);
-                    if (polyline != null) {
-                        polyline.setEndCap(new ButtCap());
-                        if (path.size() == 0) {
-                            if (destMarker != null) {
-                                LatLng destination = destMarker.getPosition();
-                                animateMarkerNew(src, destination, currMarker, eta );
-                                Polyline overlayPolylines = getPolyLine(true, polylineDataPoints);
-                                if (overlayPolylines != null) {
-                                    overlayPolylines.remove();
-                                }
-                                polyline.remove();
-                                if(polylineDataPoints != null)
-                                {
-                                    polylineDataPoints.setPolyline(null);
-                                    polylineDataPoints.setOverlayPolylines(null);
-                                }
-                                setPolyLineDataByMapInstance(pureScriptID, polylineKey, polylineDataPoints);
-                                if(currMarker != null) {currMarker.setAnchor(0.5f, 0);}
-                                mapUpdate.isMapMoved = false;
-                                mapUpdate.isMapIdle = true;
-                                animateCamera(destMarker.getPosition().latitude, destMarker.getPosition().longitude, zoomLevel, ZoomType.ZOOM);
-                            }
-                        } else {
-                                double destinationLat = path.get(0).latitude;
-                                double destinationLon = path.get(0).longitude;
-                                double sourceLat = path.get(path.size() - 1).latitude;
-                                double sourceLong = path.get(path.size() - 1).longitude;
-                                LatLng destination = path.get(path.size() - 1);
-                                animateMarkerNew(src, destination, currMarker, eta );
-                                PatternItem dash = new Dash(dashUnit);
-                                PatternItem gap = new Gap(gapUnit);
-                                List<PatternItem> PATTERN_POLYLINE_DOTTED_DASHED = Arrays.asList(dash, gap);
-                                polyline.setPattern(PATTERN_POLYLINE_DOTTED_DASHED);
-                                polyline.setPoints(path);
-                                if(debounceAnimateCameraCounter == null) debounceAnimateCameraCounter = mapRemoteConfig.debounceAnimateCameraCounter;
-                                if (autoZoom && mapUpdate.isMapIdle && (debounceAnimateCameraCounter <= 0 || mapUpdate.isMapMoved)) {
-                                    moveCamera(sourceLat, sourceLong, destinationLat, destinationLon, coordinates);
-                                    mapUpdate.isMapMoved = false;
-                                    mapUpdate.isMapIdle = true;
-                                    debounceAnimateCameraCounter = mapRemoteConfig.debounceAnimateCameraCounter;
-                                } else {
-                                    debounceAnimateCameraCounter--;
-                                }
-                            }
-                        }
-                    } catch (JSONException e) {
-                        e.printStackTrace();
-                    }
-                }
-            } catch (JSONException e) {
-                    throw new RuntimeException(e);
-                }
-        });
-    }
 
-    private void animateMarkerNew(String src, LatLng destination, final Marker marker, String infoLabelText) {
-        if (marker != null) {
-            LatLng startPosition = marker.getPosition();
-            ValueAnimator valueAnimator = ValueAnimator.ofFloat(0, 1);
-            valueAnimator.setDuration(2000); // TODO :: get this value from Loacl Storage to maintain sync with PS
-            valueAnimator.setInterpolator(new LinearInterpolator());
-            valueAnimator.addUpdateListener(animation -> {
-                try {
-                    float v = animation.getAnimatedFraction();
-                    LatLng newPosition = SphericalUtil.interpolate(startPosition, destination, v);
-                    float rotation = bearingBetweenLocations(startPosition, destination);
-                    if (rotation > 1.0){
-                        MarkerConfig markerConfig = new MarkerConfig();
-                        markerConfig.locationName(infoLabelText);
-                        markerConfig.setRotation(rotation);
-                        marker.setRotation(0.0f);
-                        marker.setIcon(BitmapDescriptorFactory.fromBitmap(getMarkerBitmapFromView(src, false, MarkerType.NORMAL_MARKER, markerConfig)));
-                    }
-                    marker.setPosition(newPosition);
-                    marker.setAnchor(0.5f, 0.7f);
-                    markers.put(src, marker);
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
-            });
-            valueAnimator.addListener(new AnimatorListenerAdapter() {
-                @Override
-                public void onAnimationEnd(Animator animation) {
-                    super.onAnimationEnd(animation);
-                }
-            });
-            valueAnimator.start();
-        }
-    }
+
 
     @JavascriptInterface
     public void removeLabelFromMarker(float zoomLevel) {
@@ -312,75 +104,6 @@ public class MobilityCustomerBridge extends MobilityCommonBridge {
     }
 
     //endregion
-
-    private float bearingBetweenLocations(LatLng latLng1, LatLng latLng2) {
-        double PI = 3.14159;
-        double lat1 = latLng1.latitude * PI / 180;
-        double long1 = latLng1.longitude * PI / 180;
-        double lat2 = latLng2.latitude * PI / 180;
-        double long2 = latLng2.longitude * PI / 180;
-        double dLon = (long2 - long1);
-        double y = Math.sin(dLon) * Math.cos(lat2);
-        double x = Math.cos(lat1) * Math.sin(lat2) - Math.sin(lat1)
-                * Math.cos(lat2) * Math.cos(dLon);
-        double brng = Math.atan2(y, x);
-        brng = Math.toDegrees(brng);
-        brng = (brng + 360) % 360;
-        return (float) brng;
-    }
-
-    @JavascriptInterface
-    public String getExtendedPath(String json) throws JSONException {
-        ArrayList<LatLng> path = new ArrayList<>();
-        ArrayList<LatLng> extendedPath = new ArrayList<>();
-        JSONObject jsonObject = new JSONObject(json);
-        JSONArray coordinates = jsonObject.getJSONArray("points");
-        if (coordinates.length() <= 1) return json;
-        int pointsFactor = Integer.parseInt(getKeysInSharedPref("POINTS_FACTOR"));
-
-        for (int i = coordinates.length() - 1; i >= 0; i--) {
-            JSONObject coordinate = (JSONObject) coordinates.get(i);
-            double lng = coordinate.getDouble("lng");
-            double lat = coordinate.getDouble("lat");
-            LatLng tempPoints = new LatLng(lat, lng);
-            path.add(tempPoints);
-        }
-        for (int i = 0, j = 1; i < path.size() - 1 && j <= path.size() - 1; i++, j++) {
-            LatLng point1 = path.get(i);
-            LatLng point2 = path.get(j);
-            extendedPath.add(point1);
-            double distanceBtw = SphericalUtil.computeDistanceBetween(point1, point2);
-            int noOfPoints = (int) Math.ceil(distanceBtw / pointsFactor);
-            float fraction = 1.0f / (noOfPoints + 1);
-            for (int k = 1; k <= noOfPoints; k++) {
-                LatLng point = getNewLatLng(fraction * k, point1, point2);
-                extendedPath.add(point);
-            }
-        }
-        extendedPath.add(path.get(path.size() - 1));
-        JSONObject newPoints = new JSONObject();
-        JSONArray remainingPoints = new JSONArray();
-        for (int i = extendedPath.size() - 1; i >= 0; i--) {
-            LatLng point = extendedPath.get(i);
-            JSONObject tempPoints = new JSONObject();
-            tempPoints.put("lat", point.latitude);
-            tempPoints.put("lng", point.longitude);
-            remainingPoints.put(tempPoints);
-        }
-        newPoints.put("points", remainingPoints);
-        return newPoints.toString();
-    }
-
-    private LatLng getNewLatLng(float fraction, LatLng a, LatLng b) {
-        double lat = (b.latitude - a.latitude) * fraction + a.latitude;
-        double lngDelta = b.longitude - a.longitude;
-        // Take the shortest path across the 180th meridian.
-        if (Math.abs(lngDelta) > 180) {
-            lngDelta -= Math.signum(lngDelta) * 360;
-        }
-        double lng = lngDelta * fraction + a.longitude;
-        return new LatLng(lat, lng);
-    }
 
     @JavascriptInterface
     public void fetchAndUpdateCurrentLocation(String callback) {
